+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## Unreleased
+### Changed
+- Rewrote example for high-level DigiROM API
+
 ## 0.7.0 - 2023-05-06
 ### Added
 - Calculation features in sequence DigiROMs: